+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Polyverse Boost ReST API (Backend)
 ### Enhancements
 - Return the owner of a project for /api/user_project/{org}/{project} (GET) - to enable user to see who owns the project
 - Return 404, 401 or 500 status code if trying to get project status for a non-existent project /api/user_project/{org}/{project}/status (GET)
+- Print a scary warning if user tries to use ENABLE_UNSIGNED_AUTHN (e.g. x-user-account) - since most APIs don't support it
+- Ensure we normalize all emails / identities including signed headers
 
 ### Bug Fixes
 - N/A